+++ src/components/_atoms/PercentageWithIcon.tsx
