+++ src/components/_atoms/RegistrationIcon.tsx
