+++ src/components/tables/nviAdmin/NviAdminPointsTableRow.tsx
@@ -1,26 +1,27 @@
 import { TableCell, TableRow } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import { PercentageWithIcon } from '../../../../components/atoms/PercentageWithIcon';
-import { CenteredTableCell, HorizontalBox } from '../../../../components/styled/Wrappers';
-import { InstitutionReport } from '../../../../types/nvi.types';
 import {
-  getNviApprovedByEverybody,
   getNviApprovedCount,
+  getNviCountApprovedByEverybody,
   getNviCountOthersMustApprove,
   getNviInstitutionName,
+  getNviPointsForReporting,
   getNviSectorLabel,
-  getNviValidPoints,
-} from './nviAdminHelpers';
+} from '../../../helpers/nviAdminHelpers';
+import { CenteredTableCell } from '../../../styles/tableStyles';
+import { InstitutionReport } from '../../../types/nvi.types';
+import { PercentageWithIcon } from '../../_atoms/PercentageWithIcon';
+import { HorizontalBox } from '../../styled/Wrappers';
 
 interface NviAdminPublicationPointsRowProps {
   report: InstitutionReport;
 }
 
-export const NviAdminPublicationPointsRow = ({ report }: NviAdminPublicationPointsRowProps) => {
+export const NviAdminPointsTableRow = ({ report }: NviAdminPublicationPointsRowProps) => {
   const { t } = useTranslation();
   const { id, institutionSummary } = report;
   const { totals } = institutionSummary;
-  const approvedByEverybody = getNviApprovedByEverybody(report);
+  const approvedByEverybody = getNviCountApprovedByEverybody(report);
   const undisputedTotals = totals.undisputedTotalCount;
   const percentageControlled = undisputedTotals > 0 ? approvedByEverybody / undisputedTotals : 0;
 
@@ -31,7 +32,7 @@ export const NviAdminPublicationPointsRow = ({ report }: NviAdminPublicationPoin
       <CenteredTableCell>{getNviApprovedCount(report)}</CenteredTableCell>
       <CenteredTableCell>{getNviCountOthersMustApprove(report)}</CenteredTableCell>
       <CenteredTableCell>{approvedByEverybody}</CenteredTableCell>
-      <CenteredTableCell>{getNviValidPoints(report)}</CenteredTableCell>
+      <CenteredTableCell>{getNviPointsForReporting(report)}</CenteredTableCell>
       <TableCell>
         <HorizontalBox sx={{ justifyContent: 'center' }}>
           <PercentageWithIcon displayPercentage={Math.floor(percentageControlled * 100)} alternativeIfZero={'-'} />
