+++ src/pages/messages/TasksPage.tsx
@@ -4,6 +4,7 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 import { Navigate, Outlet, Route, Routes, useLocation, useNavigate } from 'react-router';
+import { useFetchTickets } from '../../api/hooks/useFetchTickets';
 import { useFetchUserQuery } from '../../api/hooks/useFetchUserQuery';
 import { NviCandidateGlobalStatusEnum, NviCandidateStatusEnum, TicketSearchParam } from '../../api/searchApi';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
@@ -19,24 +20,23 @@ import { dataTestId } from '../../utils/dataTestIds';
 import { PrivateRoute } from '../../utils/routes/Routes';
 import { resetPaginationAndNavigate } from '../../utils/searchHelpers';
 import { getNviCandidatesSearchPath, getSubUrl, UrlPathTemplate } from '../../utils/urlPaths';
+import { checkUserRoles } from '../../utils/user-helpers';
 import { PortfolioSearchPage } from '../editor/PortfolioSearchPage';
 import NotFound from '../errorpages/NotFound';
 import { RegistrationLandingPage } from '../public_registration/RegistrationLandingPage';
+import { NviPublicationPointsPage } from '../tasks/nvi/NviPublicationPointsPage';
 import { NviCandidatePage } from './components/NviCandidatePage';
 import { NviCandidatesList } from './components/NviCandidatesList';
 import { NviCandidatesNavigationAccordion } from './components/NviCandidatesNavigationAccordion';
 import { NviCorrectionList } from './components/NviCorrectionList';
 import { NviCorrectionListNavigationAccordion } from './components/NviCorrectionListNavigationAccordion';
+import { NviDisputePage } from './components/NviDisputePage';
 import { NviStatusPage } from './components/NviStatusPage';
 import { ResultRegistrationsNavigationListAccordion } from './components/ResultRegistrationsNavigationListAccordion';
 import { TicketList } from './components/TicketList';
 import { TicketTypeTag } from './components/TicketTypeTag';
-import { NviDisputePage } from './components/NviDisputePage';
-import { NviPublicationPointsPage } from './components/NviPublicationPointsPage';
-import { useFetchTickets } from '../../api/hooks/useFetchTickets';
-import { useGetNotificationCounts, useGetTicketsCounts } from './user-dialog-helpers';
 import { checkPages } from './tasks-helpers';
-import { checkUserRoles } from '../../utils/user-helpers';
+import { useGetNotificationCounts, useGetTicketsCounts } from './user-dialog-helpers';
 
 const TasksPage = () => {
   const { t } = useTranslation();
