+++ src/pages/messages/components/NviCandidatesList.tsx
@@ -17,12 +17,12 @@ import { RoleName } from '../../../types/user.types';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { useNviCandidatesParams } from '../../../utils/hooks/useNviCandidatesParams';
 import { syncParamsWithSearchFields } from '../../../utils/searchHelpers';
+import { CoPublicationsCheckbox } from './CoPublicationsCheckbox';
 import { ExcludeSubunitsCheckbox } from './ExcludeSubunitsCheckbox';
 import { NviCandidateListItem } from './NviCandidateListItem';
 import { NviSortSelector } from './NviSortSelector';
 import { NviStatusFilter } from './NviStatusFilter';
 import { NviYearSelector } from './NviYearSelector';
-import { CoPublicationsCheckbox } from './CoPublicationsCheckbox';
 
 export const NviCandidatesList = () => {
   const { t } = useTranslation();
