+++ src/pages/registration/resource_type_tab/SelectRegistrationTypeField.tsx
@@ -4,8 +4,8 @@ import { ErrorMessage, useFormikContext } from 'formik';
 import { useContext, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
-import { DoesNotSupportFileIcon } from '../../../components/atoms/DoesNotSupportFileIcon';
-import { NviApplicableIcon } from '../../../components/atoms/NviApplicableIcon';
+import { DoesNotSupportFileIcon } from '../../../components/_atoms/DoesNotSupportFileIcon';
+import { NviApplicableIcon } from '../../../components/_atoms/NviApplicableIcon';
 import { CategorySelector } from '../../../components/CategorySelector';
 import { ConfirmDialog } from '../../../components/ConfirmDialog';
 import { StyledInfoBanner } from '../../../components/styled/Wrappers';
