+++ src/pages/search/project_search/ProjectSearch.tsx
@@ -7,8 +7,8 @@ import { ProjectSortSelector } from '../../../components/ProjectSortSelector';
 import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
 import { SearchParam } from '../../../utils/searchHelpers';
 import { CristinSearchPagination } from '../CristinSearchPagination';
-import { ProjectListItem } from './ProjectListItem';
 import { SearchPropTypes } from '../registration_search/RegistrationSearch';
+import { ProjectListItem } from './ProjectListItem';
 
 type ProjectSearchProps = Pick<SearchPropTypes, 'projectQuery'>;
